Change label comparisons to uint32_t as fix for host intersect; bug fix in intersect.

+++ k2/csrc/fsa_algo.cu
@@ -263,7 +263,7 @@ namespace {
 struct ArcComparer {
   __host__ __device__ __forceinline__ bool operator()(const Arc &lhs,
                                                       const Arc &rhs) const {
-    return lhs.label < rhs.label;
+    return static_cast<uint32_t>(lhs.label) < static_cast<uint32_t>(rhs.label);
   }
 };
 }  // namespace
